+++ contracts/cw-rules/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod contract;
 mod error;
 pub mod msg;
+#[cfg(test)]
+mod tests;
 
 pub use crate::error::ContractError;
