+++ contracts/cw-rules/src/tests/mod.rs
@@ -1,0 +1,1 @@
+mod nft;
